Reorganize syncer modules and classes in a new module-package
