+++ opengever/task/deadline_modifier.py
@@ -1,7 +1,7 @@
 from five import grok
 from opengever.task.browser.transitioncontroller import get_checker
 from opengever.task.interfaces import IDeadlineModifier
-from opengever.task.syncer import sync_task_response
+from opengever.task.response_syncer import sync_task_response
 from opengever.task.task import ITask
 from opengever.task.util import add_simple_response
 from zExceptions import Unauthorized
