+++ opengever/task/tests/test_response_syncer.py
@@ -4,12 +4,12 @@ from opengever.ogds.base.Extensions.plugins import activate_request_layer
 from opengever.ogds.base.interfaces import IInternalOpengeverRequestLayer
 from opengever.task.adapters import IResponseContainer
 from opengever.task.interfaces import IResponseSyncerSender
-from opengever.task.syncer import BaseResponseSyncerReceiver
-from opengever.task.syncer import BaseResponseSyncerSender
-from opengever.task.syncer import CommentResponseSyncerSender
-from opengever.task.syncer import ModifyDeadlineResponseSyncerSender
-from opengever.task.syncer import ResponseSyncerSenderException
-from opengever.task.syncer import WorkflowResponseSyncerSender
+from opengever.task.response_syncer import BaseResponseSyncerReceiver
+from opengever.task.response_syncer import BaseResponseSyncerSender
+from opengever.task.response_syncer import ResponseSyncerSenderException
+from opengever.task.response_syncer.comment import CommentResponseSyncerSender
+from opengever.task.response_syncer.deadline import ModifyDeadlineResponseSyncerSender
+from opengever.task.response_syncer.workflow import WorkflowResponseSyncerSender
 from opengever.testing import FunctionalTestCase
 from plone import api
 from plone.app.testing import TEST_USER_ID
